requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit

